+++ src/main/java/lotto/domain/exception/NotValidLottoLengthException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.domain.exception;
 
 public class NotValidLottoLengthException extends RuntimeException {
     private static final String MESSAGE = "번호는 6개를 입력해야 합니다.";
