+++ src/main/java/lotto/domain/exception/NotValidLottoTicketOfNumberException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.domain.exception;
 
 public class NotValidLottoTicketOfNumberException extends RuntimeException{
     private static final String MESSAGE = "로또를 살 금액이 부족합니다.";
