+++ src/main/java/lotto/domain/lotto/LottoManualTicket.java
@@ -1,12 +1,11 @@
-package lotto.lotto;
+package lotto.domain.lotto;
 
 import java.util.List;
 
-public class LottoAutomaticTicket extends Lotto{
-
+public class LottoManualTicket extends Lotto{
     private final List<Integer> numbers;
 
-    public LottoAutomaticTicket(List<Integer> numbers) {
+    public LottoManualTicket(List<Integer> numbers) {
         super(numbers);
         this.numbers = numbers;
     }
