+++ src/main/java/lotto/domain/lotto/LottoManualTickets.java
@@ -1,4 +1,4 @@
-package lotto.lotto;
+package lotto.domain.lotto;
 
 public class LottoManualTickets {
 }
