+++ src/test/java/lotto/PurchaseAmountTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.PurchaseAmount;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
